Added Hero Widget
Hero widget takes in a tag as a parameter. This tag is used to identify the corresponding animation on the other screen so the tag here has to match the tag on the Hero widget in the screen we are navigating to. Important to notice that the tag has to be unique. You can’t have the same tag on multiple other widgets that's why here we will use project.name as a tag as is unlikely to have another project in our portfolio with the same name,
 do the same thing in our project_detail.dart page where we will wrap our image with a Hero widget.